+++ CHANGES.rst
@@ -4,9 +4,13 @@
 4.0.10 (unreleased)
 -------------------
 
+
 - The Python implementation of ``Persistent`` and ``PickleCache`` now
   behave more similarly to the C implementation. In particular, the
   Python version can now run the complete ZODB unit test suite.
+
+- Fix the hashcode of the Python ``TimeStamp`` on 32-bit platforms.
+
 
 4.0.9 (2015-04-08)
 ------------------